# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 考我一下 API
  version: 0.1.0
  description: P0+P1 本地开发契约。开发期用 X-Dev-Openid 识别用户。
servers:
  - url: http://127.0.0.1:8000
paths:
  /health:
    get:
      summary: 健康检查
      operationId: health
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /api/home:
    get:
      summary: 首页聚合
      operationId: getHome
      parameters:
        - $ref: "#/components/parameters/DevOpenid"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeResponse"
  /api/me:
    get:
      summary: 匿名统计
      operationId: getMe
      parameters:
        - $ref: "#/components/parameters/DevOpenid"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeStats"
  /api/quiz/generate:
    post:
      summary: 提交材料并创建生成任务
      operationId: generateQuiz
      parameters:
        - $ref: "#/components/parameters/DevOpenid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResponse"
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/quiz/task/{task_id}:
    get:
      summary: 轮询任务状态
      operationId: getTask
      parameters:
        - $ref: "#/components/parameters/DevOpenid"
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/quiz/recent:
    get:
      summary: 最近闯关
      operationId: listRecent
      parameters:
        - $ref: "#/components/parameters/DevOpenid"
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, active, retest]
            default: all
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentListResponse"
  /api/quiz/{quiz_id}:
    get:
      summary: 拉取完整题目
      operationId: getQuiz
      parameters:
        - $ref: "#/components/parameters/DevOpenid"
        - name: quiz_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResponse"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/quiz/{quiz_id}/answer:
    post:
      summary: 提交单题作答
      operationId: submitAnswer
      parameters:
        - $ref: "#/components/parameters/DevOpenid"
        - name: quiz_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerResponse"
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/quiz/{quiz_id}/retest:
    post:
      summary: 基于错题生成复测
      operationId: createRetest
      parameters:
        - $ref: "#/components/parameters/DevOpenid"
        - name: quiz_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResponse"
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/question/{question_id}/feedback:
    post:
      summary: 一键纠错
      operationId: submitFeedback
      parameters:
        - $ref: "#/components/parameters/DevOpenid"
        - name: question_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /api/material/{material_id}:
    delete:
      summary: 级联删除材料与关联数据
      operationId: deleteMaterial
      parameters:
        - $ref: "#/components/parameters/DevOpenid"
        - name: material_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/events:
    post:
      summary: 前端埋点上报
      operationId: trackEvent
      parameters:
        - $ref: "#/components/parameters/DevOpenid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
components:
  parameters:
    DevOpenid:
      name: X-Dev-Openid
      in: header
      required: false
      schema:
        type: string
        default: dev-local-user
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    GenerateRequest:
      type: object
      required: [source_type]
      properties:
        source_type:
          type: string
          enum: [text, url]
        text:
          type: string
        url:
          type: string
        channel:
          type: string
    GenerateResponse:
      type: object
      required: [task_id, status]
      properties:
        task_id:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        quiz_id:
          type: string
    TaskStatus:
      type: string
      enum:
        - pending
        - extracting
        - planning
        - generating
        - validating
        - succeeded
        - degraded
        - failed
    TaskResponse:
      type: object
      required: [task_id, status, progress, stage]
      properties:
        task_id:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        progress:
          type: string
        stage:
          type: string
        quiz_id:
          type: string
        question_count:
          type: integer
        error:
          type: string
    Question:
      type: object
      required:
        - question_id
        - question_type
        - stem
        - options
        - answer_index
        - explanation
        - source_span
        - quality_status
      properties:
        question_id:
          type: string
        question_type:
          type: string
          enum: [single, true_false]
        stem:
          type: string
        options:
          type: array
          items:
            type: string
        answer_index:
          type: integer
        explanation:
          type: string
        source_span:
          type: string
        quality_status:
          type: string
          enum: [passed, degraded]
        knowledge_point:
          type: string
        ordinal:
          type: integer
    QuizResponse:
      type: object
      required: [id, material_id, question_count, is_degraded, is_retest, questions, ai_notice]
      properties:
        id:
          type: string
        material_id:
          type: string
        title:
          type: string
        question_count:
          type: integer
        is_degraded:
          type: boolean
        is_retest:
          type: boolean
        parent_quiz_id:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        attempt:
          $ref: "#/components/schemas/AttemptSummary"
        ai_notice:
          type: string
    AttemptSummary:
      type: object
      properties:
        attempt_id:
          type: string
        current_ordinal:
          type: integer
        answered_count:
          type: integer
        completed:
          type: boolean
    AnswerRequest:
      type: object
      required: [question_id, chosen_index]
      properties:
        attempt_id:
          type: string
        question_id:
          type: string
        chosen_index:
          type: integer
    QuizResult:
      type: object
      required: [correct, total, duration_seconds, wrong_question_ids]
      properties:
        correct:
          type: integer
        total:
          type: integer
        duration_seconds:
          type: integer
        wrong_question_ids:
          type: array
          items:
            type: string
    AnswerResponse:
      type: object
      required: [is_correct, correct_index, explanation, source_span, attempt_id, finished]
      properties:
        is_correct:
          type: boolean
        correct_index:
          type: integer
        explanation:
          type: string
        source_span:
          type: string
        attempt_id:
          type: string
        next_question_id:
          type: string
        finished:
          type: boolean
        result:
          $ref: "#/components/schemas/QuizResult"
    FeedbackRequest:
      type: object
      required: [error_type]
      properties:
        error_type:
          type: string
          enum: [no_evidence, wrong_answer, ambiguous, too_easy]
    EventRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          enum:
            - generation_success
            - quiz_started
            - quiz_completed
            - retest_started
            - question_error_reported
            - second_creation_7d
        payload:
          type: object
          additionalProperties: true
    Quota:
      type: object
      required: [used, limit, reset_at]
      properties:
        used:
          type: integer
        limit:
          type: integer
        reset_at:
          type: string
          format: date-time
    PrimaryTask:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [retest, continue, create]
        quiz_id:
          type: string
        material_id:
          type: string
        title:
          type: string
        wrong_count:
          type: integer
        current_ordinal:
          type: integer
        question_count:
          type: integer
    RecentItem:
      type: object
      required: [quiz_id, material_id, title, status]
      properties:
        quiz_id:
          type: string
        material_id:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [active, completed, retest]
        correct:
          type: integer
        total:
          type: integer
        wrong_count:
          type: integer
        current_ordinal:
          type: integer
        created_at:
          type: string
          format: date-time
    RecentListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecentItem"
    MeStats:
      type: object
      required: [streak_days, stars, completed_count]
      properties:
        streak_days:
          type: integer
        stars:
          type: integer
        completed_count:
          type: integer
        retest_count:
          type: integer
    HomeResponse:
      type: object
      required: [quota, primary_task, recent, me]
      properties:
        quota:
          $ref: "#/components/schemas/Quota"
        primary_task:
          $ref: "#/components/schemas/PrimaryTask"
        recent:
          type: array
          items:
            $ref: "#/components/schemas/RecentItem"
        me:
          $ref: "#/components/schemas/MeStats"
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string